requirements.txt
python-docx>=3.0.0
pytest>=7.0.0